% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{head.msdata}
\alias{head.msdata}
\title{Return ordinary data rows from an msdata object}
\usage{
\method{head}{msdata}(x, n = 6L, ...)
}
\arguments{
\item{x}{An \code{msdata} object.}

\item{n}{Number of rows to return.}

\item{...}{Ignored.}
}
\value{
An ordinary \code{data.frame} containing the first \code{n} rows.
}
\description{
Return ordinary data rows from an msdata object
}
